+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.4.0
+* Added additional filter options:
+	* Prefixing a query word or a phrase with a + (a plus) will now perform an exact search. Vendorer will attempt to match a compared value exactly and discard all other results.
+	* Range searching by item rarity e.g. >=rare will find all rare or better items.
+	* New magic words: canafford, transmogable, unknowntransmog. Transmog related filters require the dependency Can I Mog It.
+* Added a filters menu where you can choose common filtering options quickly.
+* Attempted minor performance optimizations by caching filtering results temporarily. It may or may not help.
+
 ## 2.3.4
 * Fixed paging error (again) caused by the previous fix.
 * Fixed GetMerchantItemInfo errors happening upon login.